move files from Product to Cart
