+++ api/src/App/Cart/Command/RemoveProductFromCart.php
@@ -1,20 +1,20 @@
 <?php
 
-namespace App\App\Product\Command;
+namespace App\App\Cart\Command;
 
 
 use App\Entity\Product;
 use App\Entity\Cart;
 use App\Entity\CartProduct;
 use Symfony\Component\Messenger\Attribute\AsMessageHandler;
-use App\Dto\CartItem;
+use App\Dto\CartRemoveItem;
 use App\Repository\ProductRepository;
 use App\Repository\CartRepository;
 use App\Repository\CartProductRepository;
 use DateTime;
 
 #[AsMessageHandler]
-final class AddProductToCart
+final class RemoveProductFromCart
 {
 
     public function __construct(
@@ -25,7 +25,7 @@ final class AddProductToCart
 
     }
 
-    public function __invoke(CartItem $item) : Cart | null
+    public function __invoke(CartRemoveItem $item) : Cart | null
     {
         
         $product = $this->ProductRepository->findOneById((int)$item->id);
@@ -35,32 +35,17 @@ final class AddProductToCart
         {
             
             $cartProduct = $cart->getCartProducts();
-            $cartProduct->addProduct($product);
+            $cartProduct->removeProduct($product);
             $this->CartProductRepository->update($cartProduct, true);
+            $products = $cartProduct->getProduct();
+            if (count($products) === 0)
+            {
+               $this->CartRepository->delete($cart, true);
+               return NULL;
+            }
             return $cart;
         }
-        else
-        {
-            $cart = new Cart();
-            $cart->session_id = $item->cart_id;
-            $cart->customer_id = 0;
-            $cart->total_price = 0;
-            $cart->created_at = new DateTime('now');
-            $cart->updated_at = new DateTime('now');
-            $cart->state = 'init';
-            $this->CartRepository->create($cart, true);
 
-            $cartProduct = new CartProduct();
-            $cartProduct->quantity = $item->quantity;
-            $cartProduct->addCart($cart);
-            $cartProduct->addProduct($product);
-
-            $this->CartProductRepository->create($cartProduct, true);
-            return $cart;
-
-        }
-
-        
         return NULL;
     }
 }
